+++ CHANGELOG.md
@@ -1,3 +1,15 @@
+# [1.1.0](https://github.com/kurtrank/lore/compare/v1.0.0...v1.1.0) (2025-07-01)
+
+
+### Bug Fixes
+
+* dynamically display post meta box ([#2](https://github.com/kurtrank/lore/issues/2)) ([5c78117](https://github.com/kurtrank/lore/commit/5c7811766746c61538dcf9ff39decc11a90aca1d))
+
+
+### Features
+
+* **field:** support custom label for empty item in select fields ([#3](https://github.com/kurtrank/lore/issues/3)) ([4ae8aab](https://github.com/kurtrank/lore/commit/4ae8aab4cd8c2b057562987aa908567334edd8e4))
+
 # 1.0.0 (2025-01-06)